clean up exports and files
